[WW-2734] Fix impossible casts, reported by FindBugs.
git-svn-id: https://svn.apache.org/repos/asf/struts/struts2/branches/STRUTS_2_0_X@680067 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/struts2/views/jsp/iterator/SubsetIteratorTag.java
@@ -214,7 +214,7 @@ public class SubsetIteratorTag extends StrutsBodyTagSupport {
                 count = ((Long)countObj).intValue();
             }
             else if (countObj instanceof Double) {
-                count = ((Long)countObj).intValue();
+                count = ((Double)countObj).intValue();
             }
             else if (countObj instanceof String) {
                 try {
@@ -240,7 +240,7 @@ public class SubsetIteratorTag extends StrutsBodyTagSupport {
                 start = ((Long)startObj).intValue();
             }
             else if (startObj instanceof Double) {
-                start = ((Long)startObj).intValue();
+                start = ((Double)startObj).intValue();
             }
             else if (startObj instanceof String) {
                 try {
